fix zip extraction (ignore chown errors and other warnings)
The "unzip" program prints error messages if some file in the archive
have a different owner and could not be chown'd, i.e. if walk is called
by a non-root user. It will also exit with status 1 instead of 0,
despite extracting all files. Since exit status 1 generally means
that there were some warnings, but no hard errors, we'll accept that
status as successful.