support re-entry
This is useful if a "walk" session gets interrupted
and there is still an unpacked archive directory
and an original renamed .archivename-WALK-yymmdd-hhmmss archive file.

From now on, calls like "walk archive.tar/" will succeed in such a case.
After exiting the session,
the previous renamed archive file will be deleted.